Seperated model module
